Correctif Bug x2  - En tournant vers la droite la caméra et le vaisseau faisait nimp  - Certaines toches (up down space) ne fonctionnait plus

+++ Projet/includes/constant.h
@@ -95,8 +95,8 @@ namespace gameconf
 
 namespace playercontrols
 {
-    const int NEGATIVE_MOVEMENT = -2;
-    const unsigned int POSITIVE_MOVEMENT = 2;
+    const GLfloat NEGATIVE_MOVEMENT = -2;
+    const GLfloat POSITIVE_MOVEMENT = 2;
     const GLfloat POSITIVE_SPEED_DECREASE = 0.96;
     const GLfloat NEGATIVE_SPEED_DECREASE = 0.90;
 };
